first go at the control console
The control console manages the many processes that make openreplay go.
This includes capture, the UI, and the playout backend.

This first version still has some quirks in the process management code.
It needs wxpython to work. Don't Ctrl-C it the wrong way, or you'll
end up with a mess...