Change default cut+rewind command behavior
Kyle and Reilly of RPI TV are convinced that the 100% speed replays
are in fact running at more than 100% speed. I am not sure whether
running replay of a clock next to my wristwatch convinced them
otherwise (they matched exactly).

But I do agree with them that the 100% speed replays look ugly. So
the new default is to not change speed on a cut.