Significant refactoring in ffoutput (untested)
Two ffmpeg processes are now forked - one for video, one for audio.
This eliminates the need to buffer huge quantities of data on our end. Also,
ffoutput no longer concerns itself with audio; it is simply piped direct
to aplay. This code is untested - it may not even compile. Testing today.